fix: mlx5 ib_mask_to_mlx5_opt signed 1<<31; unwind-32 hole 1017

+++ hbsd/src/sys/dev/mlx5/mlx5_ib/mlx5_ib_qp.c
@@ -2531,8 +2531,8 @@ static int ib_mask_to_mlx5_opt(int ib_mask)
 	int i;
 
 	for (i = 0; i < 8 * sizeof(int); i++) {
-		if ((1 << i) & ib_mask)
-			result |= ib_nr_to_mlx5_nr(1 << i);
+		if ((1U << i) & ib_mask)
+			result |= ib_nr_to_mlx5_nr((int)(1U << i));
 	}
 
 	return result;
